+++ composeApp/src/androidMain/kotlin/com/eddiapps/foodify/domain/AnthropicInterface.kt
@@ -3,5 +3,5 @@ package com.eddiapps.foodify.domain
 import com.eddiapps.foodify.domain.model.sheet.NutritionEntry
 
 interface AnthropicInterface {
-    suspend fun requestNutritionValues(query: String, weight: Int): Result<NutritionEntry>
+    suspend fun requestNutritionValues(query: String, weight: Double): Result<NutritionEntry>
 }
